refactor: Refactoring Directory Structure
