+++ app/src/main/java/edu/skku/cs/dokkang/data_models/MySubject.java
@@ -1,0 +1,50 @@
+package edu.skku.cs.dokkang.data_models;
+
+/*lecture information*/
+public class MySubject {
+    private String name;
+    private String professor;
+    private int id;
+    private String no;
+    private boolean checked = false;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getProfessor() {
+        return professor;
+    }
+
+    public void setProfessor(String professor) {
+        this.professor = professor;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getNo() {
+        return no;
+    }
+
+    public void setNo(String no) {
+        this.no = no;
+    }
+
+    public boolean getChecked() {
+        return checked;
+    }
+
+    public void setChecked(boolean checked) {
+        this.checked = checked;
+    }
+}
