+++ app/src/main/java/edu/skku/cs/dokkang/data_models/response/SignUpDataModel.java
@@ -1,7 +1,8 @@
-package edu.skku.cs.dokkang;
+package edu.skku.cs.dokkang.data_models.response;
 
-public class LoginDataModel {
+public class SignUpDataModel {
     private String username;
+    private String email;
     private String password;
 
     public String getUsername() {
@@ -10,6 +11,14 @@ public class LoginDataModel {
 
     public void setUsername(String username) {
         this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
     }
 
     public String getPassword() {
